Fixes race condition in CreateArchive

+++ Corgibytes.Freshli.Cli/Functionality/Git/GitArchive.cs
@@ -1,6 +1,8 @@
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
+using System.Threading.Tasks;
 using Corgibytes.Freshli.Cli.Exceptions;
 using Corgibytes.Freshli.Cli.Resources;
 
@@ -8,6 +10,9 @@ namespace Corgibytes.Freshli.Cli.Functionality.Git;
 
 public class GitArchive
 {
+    private static readonly Dictionary<string, Task<string>> s_gitIdsAndSourceTargets = new();
+    private static readonly object s_gitIdsAndSourceTargetsLock = new();
+
     private readonly ICachedGitSourceRepository _cachedGitSourceRepository;
 
     public GitArchive(ICachedGitSourceRepository cachedGitSourceRepository) =>
@@ -22,38 +27,67 @@ public class GitArchive
         // If it exists, make sure to empty it so we are certain we start with a clean slate.
         var gitSourceTarget = new DirectoryInfo(Path.Combine(cacheDirectory, "histories", gitSource.Id,
             gitCommitIdentifier.ToString()));
-        if (Directory.Exists(gitSourceTarget.FullName))
-        {
-            Directory.Delete(gitSourceTarget.FullName, true);
-        }
 
-        // Create the directory where we want to place the archive
-        gitSourceTarget.Create();
-        var archivePath = Path.Combine(gitSourceTarget.FullName, "archive.zip");
-
-        var archiveProcess = new Process
+        Task<string>? createTask = null;
+        lock (s_gitIdsAndSourceTargetsLock)
         {
-            StartInfo = new ProcessStartInfo
+            if (s_gitIdsAndSourceTargets.ContainsKey(gitSource.Id))
             {
-                FileName = gitPath,
-                WorkingDirectory = gitSource.LocalPath,
-                Arguments = $"archive --output={archivePath} --format=zip {gitCommitIdentifier}",
-                RedirectStandardOutput = true,
-                RedirectStandardError = true
+                createTask = s_gitIdsAndSourceTargets[gitSource.Id];
             }
-        };
-        archiveProcess.Start();
-        archiveProcess.WaitForExit();
-
-        if (archiveProcess.ExitCode != 0)
-        {
-            throw new GitException(string.Format(CliOutput.GitArchive_Git_Exception,
-                archiveProcess.StandardError.ReadToEnd()));
         }
 
-        ZipFile.ExtractToDirectory($"{archivePath}", gitSourceTarget.FullName);
-        File.Delete($"{archivePath}");
+        if (createTask != null)
+        {
+            return createTask.Result;
+        }
 
-        return gitSourceTarget.FullName;
+        lock (s_gitIdsAndSourceTargetsLock)
+        {
+            s_gitIdsAndSourceTargets.Add(gitSource.Id, new Task<string>(() =>
+            {
+                if (Directory.Exists(gitSourceTarget.FullName))
+                {
+                    return gitSourceTarget.FullName;
+                }
+
+                // Create the directory where we want to place the archive
+                gitSourceTarget.Create();
+                var archivePath = Path.Combine(gitSourceTarget.FullName, "archive.zip");
+
+                var archiveProcess = new Process
+                {
+                    StartInfo = new ProcessStartInfo
+                    {
+                        FileName = gitPath,
+                        WorkingDirectory = gitSource.LocalPath,
+                        Arguments = $"archive --output={archivePath} --format=zip {gitCommitIdentifier}",
+                        RedirectStandardOutput = true,
+                        RedirectStandardError = true
+                    }
+                };
+                archiveProcess.Start();
+                archiveProcess.WaitForExit();
+
+                if (archiveProcess.ExitCode != 0)
+                {
+                    throw new GitException(string.Format(CliOutput.GitArchive_Git_Exception,
+                        archiveProcess.StandardError.ReadToEnd()));
+                }
+
+                ZipFile.ExtractToDirectory($"{archivePath}", gitSourceTarget.FullName);
+                File.Delete($"{archivePath}");
+
+                return gitSourceTarget.FullName;
+            }));
+        }
+
+        lock (s_gitIdsAndSourceTargetsLock)
+        {
+            createTask = s_gitIdsAndSourceTargets[gitSource.Id];
+            createTask.Start();
+        }
+
+        return createTask.Result;
     }
 }
